// src/app/modules/carModel/carModel.interface.ts
export type TCarModel = {
    images: string[],
    brand: string,
    model: string
    variant: string
    color: string
    fuelType: string
    gearBox: string
    engineSize: string
    enginePerformance: string
    co2Emission: string
    fuelConsumption: string
    euroStandard: string
    numberPlates: string
}